desktop-sync/requirements.txt
altgraph==0.17.5
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
packaging==26.0
pillow==12.1.1
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pystray==0.19.5
python-xlib==0.33
requests==2.32.5
setuptools==82.0.0
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.5
